# Removing dependency to git2