--- a/Cargo.toml
+++ b/Cargo.toml
@@ -38,7 +38,7 @@ test = true
 default = ["build", "git", "validation"]
 build = ["dep:tonic-build", "dep:protoc"]
 validation = ["dep:anyhow", "dep:protobuf", "dep:protobuf-parse", "dep:diff-struct"]
-git = ["dep:git2"]
+git = []
 
 [dependencies]
 async-recursion = "1.0.5"
@@ -47,7 +47,6 @@ bytes = "1.0"
 clap = { version = "4.3", features = ["cargo", "derive"] }
 diff-struct = { version = "0.5.3", optional = true }
 flate2 = "1"
-git2 = { version = "0.18.1", features = ["vendored-openssl", "vendored-libgit2"], optional = true }
 hex = "0.4.3"
 home = "0.5.5"
 human-panic = "1"
@@ -75,6 +74,8 @@ assert_cmd = "2.0"
 assert_fs = "1.0"
 axum = { version = "0.7.2", default-features = false, features = ["tokio", "http1"] }
 fs_extra = "1.3"
+gix-tempfile = "13.1.0"
+gix = { version = "0.58.0", default-features = false}
 hex = "0.4.3"
 paste = "1.0.14"
 predicates = "3.0"