--- a/Cargo.toml
+++ b/Cargo.toml
@@ -74,7 +74,6 @@ assert_cmd = "2.0"
 assert_fs = "1.0"
 axum = { version = "0.7.2", default-features = false, features = ["tokio", "http1"] }
 fs_extra = "1.3"
-gix-tempfile = "13.1.0"
 gix = { version = "0.58.0", default-features = false}
 hex = "0.4.3"
 paste = "1.0.14"