--- a/tests/cmd/publish/stdout.log
+++ b/tests/cmd/publish/stdout.log
@@ -1,2 +1,4 @@
+`git status` returned an error: fatal: not a git repository (or any of the parent directories): .git
+
 :: packaged lib@0.0.1
 :: published my-repository/lib@0.0.1