This is related to https://github.com/helsing-ai/buffrs/issues/217

git2 is replaced by:
- in buffrs itself, we now call `git status --porcelain` directly, since `gix` is not able to give us information about untracked files.
- in the test-suite, we use `gix` instead of `git2`

The test suite has been updated to deal with some minor behaviour.

I run `time cargo build` before/after to compare since build time was an issue, here are the results on my laptop (M3):

- before:
```
    Finished dev [unoptimized + debuginfo] target(s) in 57.88s

________________________________________________________
Executed in   57.95 secs    fish           external
   usr time  185.65 secs    0.16 millis  185.65 secs
   sys time   55.63 secs    1.15 millis   55.63 secs
```

 - after:
 ```

    Finished dev [unoptimized + debuginfo] target(s) in 15.59s

________________________________________________________
Executed in   15.66 secs    fish           external
   usr time  102.08 secs    0.18 millis  102.08 secs
   sys time   12.89 secs    1.19 millis   12.89 secs
```