@mara-schulke 2024-02-08
Hi, is it possible to use gix for this? If not we can remove the feature altogether. Parsing cli outputs is super error prone and I would like to avoid this.