@vsiles 2024-02-09
As I mention in the issue discussion, `gix` does not yet allow us to do what `git2` was doing: `gix` does not currently have information about untracked files, so we could only complain if some tracked files have changes.
That being said, I took care of calling `git status --porcelain` because it is made exactly for that:

> --porcelain[=<version>]
> Give the output in an easy-to-parse format for scripts. This is similar to the short output, but will remain stable across Git > versions and regardless of user configuration. See below for details.

That leaves as multiple options:

1.  we remove this check altogether
2. we use `gix` and accept to ignore untracked file (for now at least)
3. we use `git` but don't parse the output. If the output is empty, the repo is clean, otherwise it is not and we print the output as is
4. we trust the `--porcelain` option and parse the output

I don't have a strong opinion here, so I'll implement what you think is best