@tomkarw 2024-02-09
I agree that using `git status --porcelain` will be robust enough, so I vote for option 4.

That said, I'm fuzzy on why we care about uncommitted changes in the first place. Maybe it's a chance to skim some overreaching features fat.