@mara-schulke 2024-02-12
Okay, so im okay with option 4 *but*: I would not error when the parsing fails. So I would at most emit a warning here that git could not be found / the output is malformed. Currently your implementation raises errors which would potentially abort the call.

If you make it best-effort (say: skip this and emit a warning line to the user) im okay with using `--porcelain`! 😊